+++ src/Home.js
@@ -1,12 +1,12 @@
 // list of all recipes in the database
 import React from "react";
 import { Link } from "react-router-dom";
+import Form from "./Form";
 
 class Home extends React.Component {
   constructor() {
     super();
     this.handleClick = this.handleClick.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
   }
 
   handleClick(event) {
@@ -15,20 +15,17 @@ class Home extends React.Component {
     this.props.removeRecipe(event.target.id);
   }
 
-  handleSubmit(event) {
-    event.preventDefault();
-    let newRecipeObject = {
-      title: event.target.title.value,
-      imageURL: event.target.imageURL.value,
-      prepTime: event.target.prepTime.value,
-      ingredients: event.target.ingredients.value,
-      story: event.target.story.value,
-      directions: event.target.directions.value,
-    };
-    this.props.addRecipe(newRecipeObject);
-  }
+  
 
   render() {
+    let defaultRecipeInput = {
+      title: "",
+      imageURL: "",
+      prepTime: "",
+      ingredients: "",
+      story: "",
+      directions: "",
+    }
     return (
       <div>
         <p>Home</p>
@@ -50,32 +47,7 @@ class Home extends React.Component {
             );
           })}
         </ol>
-        <div>
-          <h2>Add Your Own Recipe!</h2>
-          <form onSubmit={this.handleSubmit}>
-            <label htmlFor="title">Title: </label>
-            <input type="text" id="title" />
-            <br />
-            <label htmlFor="imageURL">Image URL: </label>
-            <input type="text" id="imageURL" />
-            <br />
-            <label htmlFor="prepTime">Prep Time: </label>
-            <input type="text" id="prepTime" />
-            <br />
-            <label htmlFor="ingredients">Ingredients: </label>
-            <textarea type="text" id="ingredients" />
-            <br />
-            <label htmlFor="story">Story: </label>
-            <textarea type="text" id="story" />
-            <br />
-            <label htmlFor="directions">Directions: </label>
-            <textarea type="text" id="directions" />
-            <br />
-            <button type="submit" className="btn btn-prim">
-              Submit!
-            </button>
-          </form>
-        </div>
+        <Form addRecipe={this.props.addRecipe} defaultRecipeInput={defaultRecipeInput}/>
       </div>
     );
   }
